Fix dead overflow guard in ModbusSlave frame buffer (#8)
_frameLen was uint8_t, so `_frameLen < sizeof(_frame)` (256) was always
true — the intended discard-on-overflow branch was unreachable, and a
noisy RS485 line sending 256+ bytes without a 3.5-char gap would wrap
_frameLen 255->0 and silently start overwriting the frame buffer from
index 0 instead of discarding the excess bytes. Widening to uint16_t
makes the bounds check actually take effect.

Found by compiling the firmware source against hand-written stub
headers for the Arduino/Wire/BLEDevice APIs with g++ -Wall -Wextra
-Wshadow, since PlatformIO's registry (needed for a real ESP32 cross-
compile) is blocked by this sandbox's network policy. All six source
files now compile warning-free against the stubs, and link + run
cleanly through setup()/loop(). This is not a substitute for a real
target build/hardware test, which still hasn't been done.

+++ firmware/src/ModbusSlave.h
@@ -28,7 +28,7 @@ private:
     uint16_t _vibrationCursor = 0;
 
     uint8_t _frame[256];
-    uint8_t _frameLen = 0;
+    uint16_t _frameLen = 0;  // uint16_t de propósito: precisa exceder 255 pra comparação com sizeof(_frame) ter efeito
     uint32_t _lastByteMs = 0;
 
     void handleFrame();
